requirements.txt
flask
python-pptx
matplotlib
gunicorn